revert local schema caching and rework field injection
Caching the schema locally caused too much of a performance hit in testing and memcached wasn't storing the api schema on the backend because it's too big. Pulled that out for now and reworked the field injection to iterate through records to find ones that need additional client-side values inserted.