add support for paging in find() and remove last reference to api3_preview
the 'page' parameter addition to the find() method allows the user to request a specific page of results for their query and get them back immediately instead of waiting for all pages of results to be collected before they are returned. 

the api3_preview reference was kept around for backwards compatibility for users on older versions of Shotgun. The time has come. :)