don't modify the data dict param always return thumbnail url when uploaded instead of id.
data parameter shouldn't be modified since it's a reference to the dict passed in from the client. A deep copy isn't necessary so just use copy().

thumbnail ids are pretty useless. urls are more standard so let's just return those. Since we're providing a thumbnail in create()/update() return data for them automatically to match the behavior of other fields.