Clearning password in favour of session id upon construction.
If a shotgun API instance is constructed using a username/password
combination, a session token is requested from the server and this
session token is used instead of the username/password to authenticate
for the remainder of the session. Furthermore, the password member
variable is cleared, meaning that there is no password string persisting
in memory during the lifetime of the API instance.