#30199 Update httplib2 to latest version v0.9.1 This fixes an issue where a CertificateHostnameMismatch exception would be thrown incorrectly when creating a secure connection via the API. The bug was not properly iterating through the entire list of valid hostnames to match the request against. Instead, it was only checking the first host in the list.
Updated the API to specifically set proxy_info=None when no proxy information is set in the config. Previously httplib2 defaulted to using None if this parameter wasn't provided but this changed and it now assigns a ProxyInfo instance by default. Setting this to None ensures no proxy is attempted for the connection.

Other various bug fixes specific to the library are also included. See https://github.com/jcgregorio/httplib2/blob/master/CHANGELOG for the full list.