#30340 fix unittest and possible mimetypes errors on Win Python v2.7.8
Fixes a unittest issue where the path of some tests would be incorrectly extracted. Closes #74.
Fixes a `TypeError` that could arise on Windows using Python v2.7.7/8 if you have NULL bytes in your registry. This is due to a bug in Python (http://bugs.python.org/issue23371) which was addressed in v2.7.9. The fix catches the exception that would be raised and loads the bundled and patched mimetypes library instead. Closes #88

Thanks to @patrickwolf for the reports and info