for #39129: catch empty files before uploading() and raise exception
It was possible to attempt to upload a zero-byte file which can cause all sorts of issues but most notably it generates a ServerError which is misleading to the user. We now check the file exists and is not empty before calling upload(). If the file is empty, and exception is now raised stating the issue.