#45355 - Remove filmstrip_image download test
Originally, there was a bug and the filmstrip_image url returned was bogus but it was a signed url which returned 200 and image/jpeg.
This was fixed and now, in NFS mode only, filmstrip_image returns a valid redirect url on the filmstrip_image.
That url isn't a signed url and requires login so we can't download and validate status and content-type anymore which is why this test is removed.
Note that this is not an issue in s3 mode since the filmstrip_image s3 url is returned.