SG-4483: Fixes a problem with a test fixture causing pip install problems on Windows. (#180)
There was a file included in the tests used the check compatibility with a non-standard string encoding and non-ascii characters in file names that caused problems with pip install on Windows. The file has been removed from the repo, and is now created by the relevant test on the fly.