Implement schema_field_create/update/delete
We implement the create and update using a new internal helper called `_set_property`.
This new method will recursively search the passed current property dictionary and attempt to find all keys matching the key you pass in and update it's value to the value you have passed in.

If it cannot find the value in the current dictionary but it does find a "properties" dictionary, it will add a new entry in the "properties" dictionary.
This "properties" dictionary appears to only exist on the root of the properties dictionary.

A `schema_field_update` does a deep copy of the properties before updating them whereas the `schema_field_create` method starts from a base schema_field dictionary and updates it with what you pass in.

This does not deal with setting "editable" values in each of these keys. We are assuming, in Mockgun, that you are allowed to change anything in the schema.
Perhaps someone from the team at Shotgun will have some clear logic that will allow us to implement that.