Fixed failed tests
- Fixed failed tests caused by the auto-crlf on git on Windows
- Fixed 'sudo_as_login' fails test:
It looks like the server is returning a different error than the one the test expects.  The error the server is returning seems correct for what's happening, So Will thinks that something changed SG-side and we need to update the test to match.
It's looking for an error that says The user does not have permission to 'sudo':
Instead it's seeing Cannot 'sudo' - unknown or retired user:
Which, given that blah isn't a real user, makes sense